Add files via upload
This is version one of a script that I've tested in Ubuntu 24.04 LTS Desktop.  Conducts removal of Docker and LibreOffice, if they were present, then performs a clean install of VIM, CURL, Docker and LibreOffice.  GNOME-Tweaks install is also possible, but commented out.

If for some reason a re-image needed to occur of Ubuntu 24.04 LTS Desktop, this script will handle the immediate packages that I commonly use.  Maybe it is helpful to others.