Add files via upload
Changes made in this script that differ from Version 1 are as follows:

COMMAND OUTPUT REDIRECTION AGGREGATION: In Version 1 of this script (ubuntu_fresh_v1.sh), multiple commands were executed and at the end of each execution, output redirection was performed.  Change was made to group all relative commands into braces and, in turn, take those outputs and redirect one time per grouping.  So, rather than redirecting after each command, all commands for a particular section are executed and their outputs are redirected once (after they have all been executed).

DOCKER REMOVAL: In Version 1 of this script (ubuntu_fresh_v1.sh), the Docker removal section spelled out, line-by-line, each package.  This version (Version 2) uses the wildcard (*) for Docker in the "DOCKER REMOVAL and INSTALL" section to uninstall Docker packages, which consolidates the code by several lines.

DOCKER USER: There is a docker user added in the "DOCKER REMOVAL and INSTALL" section.  This can be edited as necessary by the user.  Password is not set for user during their addition, so the user needs to set a password for the user.

ECHO: In Version 1 of this script (ubuntu_fresh_v1.sh), "echo" was used abundantly.  While it made for a more readable and cosmetically appealing log file, it took up a lot of space in the script.  "echo" has been significantly reduced in this script.

GLOBS: In the "LIBREOFFICE REMOVAL and INSTALL" section, there is a line of code to install the DEB packages that are extracted from the downloaded LibreOffice file.  Change was made to make sure that files containing dashes (-) are not interpreted as command flags/options.  Additionally, variables used in the "UPDATE SYSTEM and VERIFY SERVICES" section were changed to have double quotes to prevent globbing.

GNOME-TWEAKS: There is a section called "GNOME-TWEAKS INSTALL" that is entirely commented out.  This install may not be necessary for Ubuntu/Debian desktops, but has been left in the script just in case.  This install does not run by default (user needs to go in and uncomment out the appropriate lines of code).

LIBREOFFICE: In the "LIBREOFFICE REMOVAL and INSTALL" section, there is a line of code in which the directory needs to be changed into the newly downloaded and extracted LibreOffice directory.  A change was made to provide protection in case there is a failure in executing "cd" to that appropriate directory.

LOG FILES: Changed log file destination from "/tmp/" to "/var/" to allow review after reboot.  Also, changed outputs so that STDOUT goes to "fresh.log" and STDERR goes to "fresh_errors.log".  So, two separate log files; one for general command output (fresh.log) and one to capture all of the errors (fresh_errors.log).  Both located in /var/.

SPINNER FUNCTION: In Version 1 of this script (ubuntu_fresh_v1.sh), the spinner function was called everytime a removal or install was initiated and separate "kill" commands were executed at the end of those removals or installs.  This version (Version 2) starts the spinner at the beginning of the script (starts in the "VIM PACKAGE INSTALL" section) and then stores the spinner process in a variable.  THEN, at the end, the spinner variable (spinner process, really) gets killed.  This removed several lines of unneccessary code.

SSH SERVICE VERIFICATION: Removed this verification from the script.

VERBIAGE: Changed verbiage in both comments and statements echoed to users on terminal to be more "standard" and simple.  The description of this script was also modified.

VARIABLES: Removed the "$bold" and $normal" variables.  They ultimately were not needed.