Add files via upload
This is a script that will pull certain host, hardware and user information on the host.  This information will not only be presented to the user in the terminal, but is stored in a TXT file (if user wishes to have just a text output without the fancy stuff).

Tested on Ubuntu 24.04 LTS (Desktop) VM Image.

NOTE: When the user decides that they are done collecting information, this script will automatically remove files created inside of the script (like ".harvest" or ".harvest.txt") and clear the terminal screen for a clean start.