Add files via upload
A script to create test files in various formats for TEST environments.  A use-case for this is to create random files to test out a script or function on, without impacting other files.  Tested on Ubuntu LTS 24.04 Desktop.  There is also a "cleanup" function that will remove the created test files if they exist, but only in the directories mentioned in this script if they are there. If you do not want to cleanup the test files, just type "No" when prompted by the cleanup function.  Alternatively, you can comment out the cleanup function, if desired.