Add files via upload
A script to help organize files in a user's Desktop, Documents or Downloads directory.  This will provide the user with an options of what directory they would like to organize.  This has been tested on Ubuntu LTS 24.04 Desktop VM image.  Due to some repetitive tasks within this function, comments will be layed out per function.  Functions that repeat similar tasks will not have repeat comments (defer to the function that is commented).